% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{Clustering with EM for Mixed Data}
\usage{
em_clust(X, K, nIter, init = "kmeans")
}
\arguments{
\item{X}{A dataframe | matrix}

\item{K}{A number}

\item{nIter}{A number}

\item{init}{A string}
}
\value{
list(Log-likelihood value over all iterations, \code{z} predicted classes)
}
\description{
Clustering with EM for Mixed Data
}
