% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_fun.R
\name{loglik_fun}
\alias{loglik_fun}
\title{Compute loglikelihood value}
\usage{
loglik_fun(Vcat, Vqt, K, theta)
}
\arguments{
\item{Vcat}{An array}

\item{Vqt}{An array}

\item{K}{An array}

\item{theta}{A list}
}
\value{
loglik value
}
\description{
Compute loglikelihood value
}
