% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkfk_fun.R
\name{pkfk_fun}
\alias{pkfk_fun}
\title{Compute proportions for each cluster}
\usage{
pkfk_fun(Vcat, Vqt, K, theta)
}
\arguments{
\item{Vcat}{An array}

\item{Vqt}{An array}

\item{K}{An array}

\item{theta}{A list}
}
\value{
loglik value
}
\description{
Compute proportions for each cluster
}
