% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clust.R
\name{plot_clust}
\alias{plot_clust}
\title{Plot Clustering Results for Mixed Data}
\usage{
plot_clust(X, labels)
}
\arguments{
\item{X}{A dataframe | matrix}

\item{labels}{A list}
}
\value{
plot of individuals colored in predicted classes by EM algorithm
}
\description{
Plot Clustering Results for Mixed Data
}
