% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomInit.R
\name{RandomInit}
\alias{RandomInit}
\title{Random initialization for clustering with EM}
\usage{
RandomInit(Xqt, Xcat, K, N, hj)
}
\arguments{
\item{Xqt}{A matrix}

\item{Xcat}{A matrix}

\item{K}{A list}

\item{N}{A number}

\item{hj}{A list}
}
\value{
Model parameters : theta = list(mu, mat_cov, pk, alpha)
}
\description{
Random initialization for clustering with EM
}
